Favorite Activity and Some Bugs Fixed (abolfazl.beh) - 2018-11-15
1 - FavoriteActivity view holders layout changed.
2 - FavoriteActivity Some modification on download and separate downloaded from not downloaded entrances.
3 - FavoriteActivity new view holder for header added.
4 - EntrancePackageDownloader finish sub message for local notification changed.
5 - All HandlerThreads and Handlers chacked to fix bug:
    - BasketSingleton
    - ArchiveDetailActivity
    - EntranceDetailActivity
    - FavoriteActivity
6 - HomeActivity on load error empty holder not showed --> fixed
7 - EntranceDetailActivity cost information fixed
8 - ArchiveDetailActivity buyed time fixed.
9 - EntranceBuyDialog checked and worked correctly.
10 - SettingActivity delete purchased data modified.

+++ app/src/main/java/com/concough/android/concough/SettingActivity.java
@@ -1175,10 +1175,12 @@ public class SettingActivity extends BottomNavigationActivity {
             f = new File(SettingActivity.this.getFilesDir(), uniqueId);
         }
 
-        for (File fc : f.listFiles()) {
-            fc.delete();
+        if (f.exists() && f.isDirectory()) {
+            for (File fc : f.listFiles()) {
+                fc.delete();
+            }
+            boolean rd = f.delete();
         }
-        boolean rd = f.delete();
     }
 
 }
